+++ src/cacheout/__init__.py
@@ -18,4 +18,5 @@ from .memoization import (
     rr_memoize,
 )
 from .mru import MRUCache
+from .roundTTL import RoundTTL
 from .rr import RRCache